Move ec2.go from lib/utils to lib/utils/aws
As a result, lib/utils no longer has a dependency on aws-sdk-go.
